Add unique item crafting window
