+++ WPFSKillTree/Views/Crafting/BaseCraftingView.xaml
@@ -1,0 +1,84 @@
+﻿<UserControl x:Class="POESKillTree.Views.Crafting.BaseCraftingView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:POESKillTree.Views.Crafting"
+             xmlns:controls="clr-namespace:POESKillTree.Controls"
+             xmlns:viewModels="clr-namespace:POESKillTree.ViewModels.Crafting"
+             xmlns:converter="clr-namespace:POESKillTree.Utils.Converter"
+             xmlns:items="clr-namespace:POESKillTree.Model.Items"
+             mc:Ignorable="d" 
+             d:DataContext="{d:DesignInstance viewModels:CraftingViewModel}">
+    <UserControl.Resources>
+        <converter:EnumToSpacedStringConverter x:Key="EnumToSpacedStringConverter" />
+    </UserControl.Resources>
+
+    <StackPanel>
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition />
+                <ColumnDefinition />
+            </Grid.ColumnDefinitions>
+            <ComboBox Grid.Column="0"
+                              ItemsSource="{Binding GroupList}"
+                              SelectedItem="{Binding SelectedGroup}">
+                <ComboBox.ItemTemplate>
+                    <DataTemplate>
+                        <TextBlock Text="{Binding Converter={StaticResource EnumToSpacedStringConverter}}"/>
+                    </DataTemplate>
+                </ComboBox.ItemTemplate>
+                <ComboBox.Style>
+                    <Style TargetType="ComboBox" BasedOn="{StaticResource {x:Type ComboBox}}">
+                        <Style.Triggers>
+                            <DataTrigger Binding="{Binding TypeList.Count}" Value="1">
+                                <Setter Property="Grid.ColumnSpan" Value="2" />
+                            </DataTrigger>
+                        </Style.Triggers>
+                    </Style>
+                </ComboBox.Style>
+            </ComboBox>
+            <ComboBox Grid.Column="1"
+                              ItemsSource="{Binding TypeList}"
+                              SelectedItem="{Binding SelectedType}">
+                <ComboBox.ItemTemplate>
+                    <DataTemplate>
+                        <TextBlock Text="{Binding Converter={StaticResource EnumToSpacedStringConverter}}"/>
+                    </DataTemplate>
+                </ComboBox.ItemTemplate>
+                <ComboBox.Style>
+                    <Style TargetType="ComboBox" BasedOn="{StaticResource {x:Type ComboBox}}">
+                        <Style.Triggers>
+                            <DataTrigger Binding="{Binding TypeList.Count}" Value="1">
+                                <Setter Property="Visibility" Value="Collapsed" />
+                            </DataTrigger>
+                        </Style.Triggers>
+                    </Style>
+                </ComboBox.Style>
+            </ComboBox>
+        </Grid>
+        <controls:SearchableComboBox ItemsSource="{Binding BaseList}"
+                                             SelectedItem="{Binding SelectedBase}">
+            <ComboBox.ItemTemplate>
+                <DataTemplate DataType="{x:Type items:ItemBase}">
+                    <StackPanel Orientation="Horizontal">
+                        <Image Height="20" VerticalAlignment="Center" Source="{Binding Image.ImageSource.Result}"
+                                       StretchDirection="DownOnly" Margin="0,0,3,0"/>
+                        <TextBlock VerticalAlignment="Center">
+                            <TextBlock.Text>
+                                <MultiBinding StringFormat="{}{0} (ilvl {1})">
+                                    <Binding Path="Name" />
+                                    <Binding Path="Level" />
+                                </MultiBinding>
+                            </TextBlock.Text>
+                        </TextBlock>
+                    </StackPanel>
+                </DataTemplate>
+            </ComboBox.ItemTemplate>
+        </controls:SearchableComboBox>
+        <local:ModSelectorView Height="25"
+                                       DataContext="{Binding MsQuality}" />
+        <local:ModSelectorView Height="40"
+                                       DataContext="{Binding MsImplicits}" />
+    </StackPanel>
+</UserControl>
