+++ WPFSKillTree/Views/Crafting/BaseCraftingView.xaml.cs
@@ -1,0 +1,13 @@
+﻿namespace POESKillTree.Views.Crafting
+{
+    /// <summary>
+    /// Interaction logic for BaseCraftingView.xaml
+    /// </summary>
+    public partial class BaseCraftingView
+    {
+        public BaseCraftingView()
+        {
+            InitializeComponent();
+        }
+    }
+}
