+++ WPFSKillTree/Views/Crafting/UniqueCraftingView.xaml
@@ -1,0 +1,68 @@
+﻿<dialogs:CloseableBaseDialog x:Class="POESKillTree.Views.Crafting.UniqueCraftingView"
+                             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+                             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+                             xmlns:local="clr-namespace:POESKillTree.Views.Crafting"
+                             xmlns:dialogs="clr-namespace:POESKillTree.Controls.Dialogs"
+                             xmlns:crafting="clr-namespace:POESKillTree.ViewModels.Crafting"
+                             xmlns:l="clr-namespace:POESKillTree.Localization.XAML"
+                             xmlns:controls="clr-namespace:POESKillTree.Controls"
+                             mc:Ignorable="d"
+                             d:DataContext="{d:DesignInstance crafting:UniqueCraftingViewModel}"
+                             CloseButtonVisibility="Collapsed">
+    <dialogs:CloseableBaseDialog.Resources>
+        <crafting:UniqueCraftingViewModelProxy x:Key="VmProxy" Data="{Binding}" />
+    </dialogs:CloseableBaseDialog.Resources>
+    
+    <dialogs:CloseableBaseDialog.Title>
+        <l:Catalog Message="Unique Crafting Window"/>
+    </dialogs:CloseableBaseDialog.Title>
+    <dialogs:CloseableBaseDialog.DialogLeft>
+        <controls:ItemTooltip HorizontalAlignment="Center" VerticalAlignment="Top"
+                              Margin="10 52"
+                              Item="{Binding Item}" />
+    </dialogs:CloseableBaseDialog.DialogLeft>
+
+    <Grid MinWidth="400" Margin="0 5 0 0">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+
+        <local:BaseCraftingView Grid.Row="0"
+                                DataContext="{Binding}" />
+
+        <Grid Grid.Row="1"
+              Margin="0 5 0 0">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="*"/>
+                <ColumnDefinition Width="auto"/>
+            </Grid.ColumnDefinitions>
+
+            <local:ModSelectorView Grid.Column="0"
+                                   DataContext="{Binding MsExplicits}" />
+
+            <controls:ItemVisualizer Grid.Column="1"
+                                     Width="94" Height="188"
+                                     Margin="5"
+                                     VerticalAlignment="Top"
+                                     Item="{Binding Data.Item, Source={StaticResource VmProxy}}"/>
+        </Grid>
+
+        <StackPanel Grid.Row="2"
+                    Style="{StaticResource DialogPanel}">
+            <Button Style="{StaticResource DynamicDialogPanelButton}"
+                    Command="{Binding CloseCommand}" CommandParameter="{StaticResource True}"
+                    IsDefault="True">
+                <l:Catalog Message="Create Item"/>
+            </Button>
+            <Button Style="{StaticResource DynamicDialogPanelButton}"
+                    Command="{Binding CloseCommand}" CommandParameter="{StaticResource False}"
+                    IsCancel="True">
+                <l:Catalog Message="Cancel"/>
+            </Button>
+        </StackPanel>
+    </Grid>
+</dialogs:CloseableBaseDialog>
