+++ WPFSKillTree/Views/Crafting/UniqueCraftingView.xaml.cs
@@ -1,0 +1,13 @@
+﻿namespace POESKillTree.Views.Crafting
+{
+    /// <summary>
+    /// Interaction logic for UniqueCraftingView.xaml
+    /// </summary>
+    public partial class UniqueCraftingView
+    {
+        public UniqueCraftingView()
+        {
+            InitializeComponent();
+        }
+    }
+}
