+++ WPFSKillTree/Views/MainWindow.xaml
@@ -704,7 +704,7 @@
                                        Height="2" Fill="{DynamicResource GrayNormalBrush}" VerticalAlignment="Top" />
                             
                             <UniformGrid Grid.Column="0" Grid.Row="0"
-                                         Margin="0 6 62 0"
+                                         Margin="0 6 60 0"
                                          Height="32" Width="480"
                                          Rows="1"
                                          HorizontalAlignment="Right">
@@ -715,6 +715,10 @@
                                 <Button Margin="4 0"
                                         Click="Menu_ImportStash">
                                     <l:Catalog>Import Stash</l:Catalog>
+                                </Button>
+                                <Button Margin="4 0"
+                                        Click="Button_CraftUnique_Click">
+                                    <l:Catalog Message="Craft new unique"/>
                                 </Button>
                                 <Button Margin="4 0"
                                         Click="Button_Craft_Click">
