+++ WPFSKillTree/WPFSKillTree.csproj
@@ -209,10 +209,16 @@
       <DependentUpon>ColorPickerButton.xaml</DependentUpon>
     </Compile>
     <Compile Include="Controls\SearchableComboBox.cs" />
+    <Compile Include="Model\Items\IItemBase.cs" />
+    <Compile Include="ViewModels\Crafting\AbstractCraftingViewModel.cs" />
     <Compile Include="ViewModels\Crafting\CraftingViewModel.cs" />
     <Compile Include="ViewModels\Crafting\ModSelectorViewModel.cs" />
     <Compile Include="ViewModels\Crafting\SliderGroupViewModel.cs" />
     <Compile Include="ViewModels\Crafting\SliderViewModel.cs" />
+    <Compile Include="ViewModels\Crafting\UniqueCraftingViewModel.cs" />
+    <Compile Include="Views\Crafting\BaseCraftingView.xaml.cs">
+      <DependentUpon>BaseCraftingView.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Views\Crafting\CraftingView.xaml.cs">
       <DependentUpon>CraftingView.xaml</DependentUpon>
     </Compile>
@@ -487,6 +493,9 @@
     <Compile Include="Views\BuildsControl.xaml.cs">
       <DependentUpon>BuildsControl.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Views\Crafting\UniqueCraftingView.xaml.cs">
+      <DependentUpon>UniqueCraftingView.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Views\DownloadStashWindow.xaml.cs">
       <DependentUpon>DownloadStashWindow.xaml</DependentUpon>
     </Compile>
@@ -533,6 +542,10 @@
     <Page Include="Controls\ColorPickerButton.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Views\Crafting\BaseCraftingView.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Views\Crafting\CraftingView.xaml">
       <SubType>Designer</SubType>
@@ -633,6 +646,10 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Views\BuildsControl.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Views\Crafting\UniqueCraftingView.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
